+++ terraform/deployment/default.nix
@@ -5,6 +5,9 @@ in
 attrs@{ terraformPackage, preDeployPhase ? "", postDeployPhase ? "", shellInputs ? [ ], inputs ? [ ], ... }:
 base.deployment.mkDeployment (attrs // {
   name = "deploy-terraform-${terraformPackage.name}";
-  deployPhase = ''command ${deployer}/bin/terraform-deploy --source ${terraformPackage}/src "$@"'';
+  deployPhase = ''
+    set -e
+    command ${deployer}/bin/terraform-deploy --source ${terraformPackage}/src "$@"
+  '';
   inherit preDeployPhase postDeployPhase shellInputs inputs;
 })